feat(API): add an endpoint to synchronize the local cart with the API
This commit introduces an endpoint that allows users to send their
local cart data and persist it to the API. Anonymous users cannot add
cart products via API requests, so they need to store them in
localStorage. If a user authenticates, they can send a request to this
endpoint, specifying their local cart data, which will then be saved in
the database.

For a more detailed explanation of this endpoint please refer to the
API documentation.

This endpoint is needed for the React app to seamlessly persist a
user's local cart data to the backend upon user authentication.