feat(React): add the ability for anonymous users to interact with the cart
Previously, anonymous users weren't able to interact with the cart due
to a fact that the API doesn't accept cart requests from anonymous
users. But now that the API server includes an endpoint that allows to
synchronize local cart products with the API, and another endpoint for
checking if it's safe to add a product to the local cart (if there are
enough products in stock and the limits aren't exceeded), implementing
cart logic for anonymous users became possible using client-side logic
with React. Local cart data is stored in localStorage, but interaction
with the cart happens via Redux. When the React app initially loads,
it grabs cart data from localStorage, and then stores it in the Redux
store. When a user adds cart products, modifies their quantity, or
completely deletes them, the React app changes the Redux store via
reducers, then these reducers change the state, and then a middleware
is run that persists this data to localStorage.

It's also important to mention that when working with the cart locally,
API requests are still made, for example to check whether there are
enough products in stock, or to check how many cart products a single
user is allowed to have. Additionally, when a user logs in, their local
cart is synchronized with the API. Which products are synchronized is
determined by the API logic.