feat(API): add 59 more users to the DB using seeding files
This commit modifies the '02_users.ts' seeding file to include 59 more
users. Overall, there will be 60 users in the application. This is
needed to add a seeding file for product reviews in the future. It
won't make much sense if only 1 user writes all the reviews in the
application.

In addition to this, strong passwords for these accounts are generated
automatically and are hashed appropriately.