feat(API): add a migration file that creates the 'product_review_comments' table
This table stores replies from users to product reviews.

Simplified table structure:
- id SERIAL
- review_product_id - foreign key pointing at the composite primary key
of the 'product_reviews' table
- review_owner_id - foreign key pointing at the composite primary key
of the 'product_reviews' table
- comment_owner_id - id of a user who wrote the comment to a product
review
- comment_message - VARCHAR(2000)
- moderation_status_id - foreign key pointing at 'moderation_statuses'
- created_at - DATE (stores year, month and day, no time of day)

This table won't be used in the recent future, because implementing a
feature to add replies to user reviews isn't a priority right now. But
it will be added at some point in the future.