fix(API documentation): rewrite the authentication lack scenario
The API documentation had this line about an error scenario in GraphQL
mutations:
"The user is not authenticated (the 'Authorization' header is empty)"

However, this is not entirely accurate. This error scenario is also
relevant if the provided access token in the 'Authorization' header is
incorrect. This commit fixes the API documentation by correcting all
mentions of this error. It's been rewritten to be like this:
"The user is not authenticated (the 'Authorization' header is empty or
the provided access token is invalid)"