refactor(React): move <ProductInfo /> to the 'pages' folder
Previously, the <ProductInfo /> component was in the 'components'
folder. However, since this component is used as a page (it's used by
React Router), it's more appropriate for this component to be in the
'pages' folder. This commit changes the location of the <ProductInfo />
component by moving it to the 'pages' folder. The functionality of the
React application wasn't affected, only the structure.