refactor(React): move error and success modals to 'components'
Previously, the <DeleteButton /> component utilized the <ErrorModal />
and <SuccessModal /> components. These 2 components were stored near
<DeleteButton />, because it was expected that these components would
only be used inside this component. However, as of now, this is not the
case. These components will be needed in the future, so this commit
lays the groundwork for future changes by moving <SuccessModal /> and
<ErrorModal /> to the 'components' folder, which stores general-purpose
components.

The functionality of the application wasn't changed at all. This commit
only performs structural changes to the React application.