feat(API GraphQL getProduct query): add the 'userCanAddReview' field
When a user fetches product information using the 'product' GraphQL
query, a new field called 'userCanAddReview' may now be included. This
field specifies whether the user who requested this query can add a
review to the specified product or not. As of now, one user can only
add one review per product. This was done intentionally to make sure
that users can't add multiple reviews to the same product. When a user
requests product information, the client application will need to know
whether to render the 'Add new review' button or not. The client app
can now just include the 'userCanAddReview' field, which will return
the needed information.

Note, however, that in order to request the 'userCanAddReview' field,
a valid access token must be provided, which means that the end user
must be authenticated.

To get more details about this query, please refer to the API
documentation, which was updated in this commit to reflect this new
change.