feat(React): add the ability to submit product reviews
Authenticated users viewing product information pages (for example
/products/729) can now see a button for adding reviews. Reviews require
a star rating (1-5 in 0.5 increments) and a message (1-2000 characters).

Upon successful submission, a confirmation message signifies that the
product review has been sent for approval. Only one review per user is
allowed for each product. If a user attempts to submit a new review
again, a modal window will be shown informing the user that adding
multiple reviews is not permitted.