fix!(API GraphQL products): enforce non-nullable results
Similarly to the 'featuredProducts' GraphQL query, the return data of
the 'products' GraphQL query can be null. This was added by accident;
the underlying GraphQL resolver never returns 'null' as the result of
the query.

This commit fixes this oversight by adjusting the return type of the
'products' query to: "GetProductsByCategoryAndPageReturnValue!". This
guarantees that the GraphQL API returns the expected type instead of
'null'.

This update simplifies the workflow for frontend developers, as the
frontend application can now accurately interpret the return type of
the 'products' GraphQL query.

It's important to note that this commit exclusively modifies backend
code. Any necessary adjustments to the frontend code may be required in
subsequent updates to align with this change.