feat(API GraphQL 'featuredProducts' query): add a 'userRating' field
The 'featuredProducts' GraphQL query returns a list of featured
products (as of now, it just fetches 12 random products).

This commit introduces a new feature that allows users to also include
the 'userRating' field, which returns how many stars each product has
depending on its reviews. The 'userRating' field ranges from 1 to 5 in
increments of 0.5.

This will allow the client application to display the rating next to
each product.