feat(React): add user rating to product details page
When a user navigates to the product details page (e.g. /products/750),
they are presented with various product information such as title,
price, and short description etc. This commit extends this page to also
display the user rating for the product, allowing users to make more
informed decisions about their purchases.

The user rating ranges from 1 to 5 in increments of 0.5.