feat(API GraphQL 'products' query): add a 'userRating' field
Similarly to the 'featuredProducts' and 'product' GraphQL queries, this
commit introduces a new 'userRating' field to the 'products' GraphQL
query. This field specifies how many stars a product has depending on
its reviews. The value can range from 1 to 5 in increments 0.5.

This will allow the client application to display the rating on the
page that displays products by category
(e.g. "/products/category/Ноутбуки?page=2")

The API documentation has also been updated to reflect this change.