feat(React): add user ratings to the page displaying products by page
When a user navigates to the page displaying products by specifying a
category and a page (e.g. /products/category/Відеокарти?page=1), they
can view various products along with a page picker. This commit extends
the functionality of this page to include the display of user ratings
for each product. The user rating ranges from 1 to 5 in increments of
0.5.