feat(API): add a migration file for 'postal_service_warehouses' table
This migration file creates a new table for storing Nova Poshta and
Ukrposhta warehouses. This table is called 'postal_service_warehouses'
and it will be used by the React application to retrieve the locations
where the ordered goods can be delivered.

This table will store these columns:
- id
- postal_service_id (references "delivery_methods"."id")
- city_id (references "cities"."id")
- warehouse_description - VARCHAR(200) - specifies a short description
of the warehouse