feat(API): add seed file to prepopulate 'postal_service_warehouses' table
The 'postal_service_warehouses' table will be used to store all post
offices where the ordered goods can be delivered. It would be too slow
to use the Nova Poshta / Ukrposhta API every time a user wants to see
available post offices, so it's better to store these in the database
and then refresh it once a day using a CRON job or an AWS Lambda
function.