feat(API): add an endpoint that retrieves supported cities
A new endpoint (GET /api/shipping/supported-cities) was introduced by
this commit. This endpoint returns a list of cities that are supported
for delivery within the application. This list of available cities will
be displayed in the React app on the checkout page.

For more information about this endpoint please refer to the API docs.