feat(API): add endpoint to retrieve user profile
This commit implements a new endpoint (GET /api/user/profile) that
allows users to retrieve their first name, last name, and phone number.
This endpoint will be used by the React application to display the
authenticated user's profile data when they'll be trying to create a
new order.

For more information about this endpoint please refer to the API docs.