feat(React <Button />): add a new color theme
This commit implements a new color theme for the <Button> component.
The default color theme is applied by default, but if you want to
change it, you can use the 'colorTheme' prop:

<Button colorTheme={2}>Confirm the order</Button>

This change is needed for the order checkout page. A button will be
displayed, and in order for it to be in harmony with the checkout page,
it's necessary to use the new color theme.