feat(GET /api/user/cart): return if products can be ordered
When a user retrieves their cart using the GET /api/user/cart endpoint,
they will now receive one additional field for each product:
- canBeOrdered - indicates whether a current product can be ordered. A
cart product can be ordered if the specified quantity doesn't exceed
the stock balance and if the "maxOrderQuantity" limit hashn't been
exceeded (each product comes with a restriction on the maximum number
of identical instances that a single user can order)

In addition, those items in the cart that exceed the stock balance or
the maximum order limit will not be included in the total price. Users
will not be able to order these products, so there is no point in
including them in the total price.

These additions enable the React application to inform users if their
cart contains an excessive quantity of products, prompting them to
adjust accordingly.

The API documentation has also been changed to reflect this new change.

Additional notes:
Redis continues to cache DB cart retrieval results, but ensuring cached
carts remain up-to-date is crucial. Currently, mechanisms to invalidate
cached carts upon product orders or admin changes to stock quantities
or maximum order limits aren't implemented. These will be addressed in
future revisions, as endpoints for order creation and product editing
are scheduled for rewriting.