feat(API): add an endpoint for checking order feasibility
POST /api/user/order/check-feasibility
This endpoint is designed to determine whether it's feasible to place
an order for the specified products. Users provide a list of product
IDs along with their desired quantities, and the server responds with
whether these products can indeed be ordered in the requested amounts.

The primary purpose of this endpoint is to support the integrity
checking of anonymous users' shopping carts in the React application.
Since anonymous users can't interact with the cart via the API
directly, they rely on storing cart data locally using localStorage.
When they access their cart, it's crucial to inform them if some
products are unavailable. Hence, this endpoint serves to validate the
cart's contents.

Moreover, this endpoint can also be utilized when anonymous users
attempt to create an order. It helps in notifying them promptly if
certain products are unavailable at the moment, prompting them to
adjust their cart accordingly.

Please refer to the API documentation for more details about this
endpoint.