feat(React): display unavailable products in the cart
Previously, when a user added a product to the cart, it was assumed
that this product would always be in stock. However, this is not the
case, as the number of products in stock can change over time, so it is
important to check if the items in the cart can be ordered. This commit
implements this functionality. Whenever a user retrieves their cart,
the available and unavailable items will be grouped together, and the
unavailable ones will be highlighted and not included in the total
price.