feat(API migrations): only support postal services
This commit adds new migration files that adjust the DB schema so that
it supports only postal services as delivery methods. Previously, the
'orders' table was designed in a way that supports other delivery
methods besides postal services, for example courier services. However,
this application will only support postal services, at least for now.
For this reason it was decided to remodel the DB schema to adhere to
this requirement.

A list of changes that the newly created migration files perform:
- Drop the 'post_office' column from the 'orders' table (because this
column doesn't point to anything, it's a simple VARCHAR column, so the
warehouse names would be duplicated)
- Drop the 'delivery_method_id' column from the 'orders' table (because
this column is redundant)
- Add the 'delivery_warehouse_id' column to the 'orders' table. This
column points to the 'id' column of the 'postal_service_warehouses'
table. This column specifies which warehouse the given order should be
delivered to. This warehouse is automatically associated with some
postal service, which is the reason why we've dropped the
'delivery_method_id' column.