fix(POST /api/auth/oauth-callback): remove 'avatar_url'
This commit fixes a bug that prevented users from signing up by using
their Google/Facebook account. This bug was caused because the backend
application tried to insert a new user specifying the URL of their
avatar. Previously, the 'users' table had a column called 'avatar_url',
but after some time it was removed because avatars weren't used
anywhere in the application. When this column was removed, the
Node.js application wasn't adjusted to account for this change. This
commit fixes this.