feat!(API): remove 'phone_number' from the 'users' DB table
This commit creates a migration file that removes the 'phone_number'
column from the 'users' table.

It was decided to drop the option to log in with phone numbers due to
the high cost of verifying them through notifications. Going forward,
users will only be able to log in with their emails.

When users create orders, they'll provide their phone numbers, which
will then be stored in the 'orders' table. Managers will later reach
out to the provided phone number to confirm the user's order. However,
this logic has not been implemented yet.

This change affects several parts of the Node.js application:
- GET /api/user/profile - when retrieving user profile, only first name
and last name will be returned (phone number won't be returned because
it's no longer an integral part of the user).
- POST /api/auth/sign-in - phone number can no longer be used as a user
login.

Additional Note:
- There is an endpoint that allows users to create an order. However,
it will not work as expected. This endpoint is not used by the React
application and it is scheduled to be completely rewritten, so it was
decided to leave this endpoint untouched, even though it utilizes the
deleted 'phone_number' column of the 'users' table.