refactor(API): move 'generateStrongPassword' to the signup model
Previously, the 'generateStrongPassword' function resided within the
OAuth model. It was initially placed there under the assumption that it
would solely be utilized for user sign-ups via Google or Facebook
accounts. In these cases, the system generated a password using
'generateStrongPassword' and stored the password in the 'users' table.
However, the function's usage has expanded to include scenarios where
anonymous users place orders. In such instances, the backend creates an
account on behalf of the user and sends the generated password to the
provided email address. Considering these use cases, this function has
been relocated to the signup model.