refactor(API): rename 'withOAuth' to 'isActivated'
This commit modifies the 'signUpUser' function of the signup model by
renaming the 'withOAuth' parameter to 'isActivated' The 'withOAuth'
parameter was originally intended to specify whether the user's
signed-up account should be immediately activated, primarily for cases
where users signed up with Google or Facebook accounts. However, the
use of this parameter has evolved, and it is now also applicable when
anonymous users create an order.

When anonymous users create an order, the backend application creates
an account for them with an auto-generated password, which is then sent
to the provided email. In such cases, account activation is not
necessary because users would need to access their email to discover
the password.