feat(API migrations): remove 'city_id' from the 'orders' DB table
This commit removes the 'city_id' column from the 'orders' table. Prior
to this commit, the 'orders' table looked like this:
- id
- user_id
- payment_method_id
- is_paid
- status_id
- delivery_warehouse_id
- city_id

'delivery_warehouse_id' points to the 'postal_service_warehouses' table,
which in turn has these columns:
- id
- postal_service_id
- city_id (note this column!)
- warehouse_description

The problem with the current design is that both 'orders' and
'postal_service_warehouses' tables have the 'city_id' column, which can
lead to inconsistencies. This is why this commit drops this column from
the 'orders' table. Specifying 'delivery_warehouse_id' automatically
specifies the needed city.