fix(API migrations): resolve rate limit issues for Nova Poshta API
This commit addresses rate limit problems encountered with the Nova
Poshta API during the execution of the '12_postal_service_warehouses.ts'
seed file. The seed file is responsible for populating the
'postal_service_warehouses' DB table by fetching information about
branches (відділення) and pickup points (пункти) from the Nova Poshta
API.

The root cause of the issue was the previous implementation of the seed
file, which did not consider the rate limits imposed by the Nova Poshta
API to prevent excessive API requests. To overcome this, the approach to
retrieving the necessary branches and pickup points has been completely
revamped. Instead of using pagination, all warehouses are now fetched in
a single API request, and the result is processed using a readable
stream. This enables the extraction of the required warehouse
descriptions on the fly without loading the entire API response into
memory.

The process involves loading the extracted data into the
'postal_service_warehouses' DB table, followed by repeating the cycle
for each city supported by the e-commerce app. By adopting this method,
the number of API requests per city is reduced from ~10 (using
pagination) to just 1, utilizing streams for both retrieval and data
extraction. Additionally, a 1-second wait is implemented between
iterations to ensure compliance with the rate limit.