fix(API): resolve an error when adding new cart product
Previously, attempting to add a new product to the cart resulted in an
error due to a flawed SQL query with typing issues. This query is
responsible for retrieving additional product details, which wouuld then
be cached in Redis. Since this query had errors, it always failed,
preventing the product from being added to Redis while still adding it
to PostgreSQL. This bug also caused users to be able to exceed the
allowed number of products in their cart, because the limit check was
performed on Redis, not on PostgreSQL. This commit rectifies the issue.