feat(API): add an endpoint that returns order recipients
This endpoint introduces a new "GET /api/user/order/recipients" endpoint
that returns different order recipients for a single user. When a user
creates a new order, they can specify who will receive this order, and
it doesn't have to be the same person. This new endpoint will return the
recipients that were used before for order creation. If a user wants to
add a new recipient, they will use the order creation endpoint
(POST /api/user/order, which is currently work in progress) and specify
who will receive the order.

It's important to point out that this new endpoint is only available for
authenticated users. If a user is anonymous, they will just provide
their contact data (first name, last name, email, phone number etc) when
creating an order using the "POST /api/user/order" endpoint, and the API
will create a new account on their behalf, and then create the desired
order.

This commit also creates a new migration file that creates the
'order_recipients' table, which is necessary for this new endpoint. This
table has the following columns:
- id
- associated_user_id (points to "users"."id", NOT NULL)
- first_name (VARCHAR(50) NOT NULL)
- last_name (VARCHAR(50) NOT NULL)
- phone_number (VARCHAR(50) NOT NULL, must adhere to this format:
'+380-12-345-67-89')