feat(API): add 'email' to the access token payload
This commit changes the data that access tokens (JWT) store. Previously,
access tokens only stored user ID and whether the user is an admin or
not. However, when implementing the order creation endpoint, there was a
need to retrieve the user's email based on their user ID. It was
possible to retrieve their email by simply making a request to the
database. However, DB requests take time, so it's better to store users'
email in the access tokens when they sign in or when they refresh their
access tokens using refresh tokens. This way, when the API server
receives the user's access tokens, it immediately knows the user's
email. Besides, emails can't be changed using the API as of today, so
there will be no need to invalidate access tokens, because emails won't
be changed.

One might argue that storing emails in the access tokens (which are JWT
tokens) is not secure because everyone can see the payload of JWT
tokens. However, email address is not such a sensitive piece of
information. Moreover, if an access token is hijacked, there will be
much bigger problems than the exposed email address.