feat(API): move 'serializeSqlParameters' to a separate file
The 'serializeSqlParameters' function is used to create binding
parameters for SQL queries. This is used in scenarios when SQL queries
are generated dynamically.
Example:
serializeSqlParameters(2) => '($1, $2),($3, $4)'

This function used to be in the product model (src/models/product.ts) as
a non-exportable function. However, this function is now required to be
used in many other places, for example in the order model. This commit
moves this function to "src/util/serializeSqlParameters.ts". This will
increase code modularity and simplify the maintenance process in the
future.