refactor(API signupModel.signUpUser): change argument name
This function renames the 'password' parameter to 'hashedPassword' in
the 'signUpUser' function of the signup model (src/models/signup.ts).
This change will improve code readability and maintainability, because
the parameter name 'password' is ambiguous, as it doesn't specify
whether the password is hashed or not. The name 'hashedPassword' is much
clearer in that regard.