refactor(API): create the 'hashPassword' function
Previously, whenever there was a need to hash a password, controllers
used "bcryptjs.hash(plaintextPassword, 12)". This increased duplication
in the code. This way, if there was a need to change the "work factor"
of the hash function (in this case it's 12), developers would need to
modify multiple places in the code. To solve this problem, a separate
function called "hashPassword" was created inside the signup model
(since hashing passwords is used for signing users up). By using this
approach, the code has become more maintainable.