feat(API): remove "GET /api/user/order/:orderId/payment-data" endpoint
This endpoint was used to generate "data" and "signature" parameters.
These parameters were used to redirect the user to the LiqPay payment
page. However, it was decided to completely rewrite the order creation
endpoint, and the generation of "data" and "signature" will be done
inside this new version of the order creation endpoint. This will
improve the security of the order creation workflow.