refactor(API): remove 'isProductInCart' function from cart model
The 'isProductInCart' function was needed to be removed from cart model
because it wasn't used anywhere in the code. Moreover, this function
had a bug.

If a need for this function arises in the future, it will be introduced.
This commit removes this function to reduce code complexity and improve
code readability.