feat(API): rewrite the order creation endpoint
This commit completely revamps the endpoint responsible for creating
new orders (POST /api/user/order). The first version of this endpoint
was written a long time ago, and it's hasn't been used by the React
application. The database schema changed significantly since the last
release, and some business requirements were altered, for example the
support for Ukrposhta was dropped; only Nova Poshta is supported at the
moment. This endpoint was rewritten because it's expected to be used
by the React application in the upcoming updates.

This new version of the endpoint now allows both authenticated and
anonymous users to create a new order. For comprehensive details on
utilizing this endpoint, please refer to the API documentation.