feat(API): rewrite the payment callback endpoint
This commit completely revamps the "POST /api/user/order/callback"
endpoint, which processes payment results. When a user completes payment
for their order, they are redirected here. The system validates that the
request is from LiqPay, marks the order as paid, and sends an email
confirmation to the user's associated address. Managers will then call
the client to confirm the order. Additionally, an alert is sent to the
admin via Telegram to notify them of the new order.

The previous implementation of this endpoint had significant flaws. For
instance, it didn't send email notifications to users upon payment
completion. Moreover, its validation rules were less stringent. This
overhaul addresses these issues. The update is crucial because this
endpoint will soon be integrated into the React application.

For a more detailed explanation of this endpoint, please refer to the
API documentation.