feat(DEV docker compose): add Elasticsearch service
The 'docker-compose.yml' file is used to configure a local deployment
of Docker containers. This commit introduces a new service called
Elasticsearch, which is a full-text search engine.

This addition lays the groundwork for implementing full-text search
functionality for products. The vision is to allow users to input search
queries, with the React application subsequently displaying a list of
relevant products. The inclusion of Elasticsearch marks the initial
step towards achieving this goal.