fix(React): accept 'userRating' as undefined
This commit modifies the following components to make the 'userRating'
prop optional:
- ProductList
- ProductItem

This is necessary because the 'userRating' prop can sometimes be of
type 'number | null | undefined', which isn't accepted by the above
components, because they expect 'userRating' to be 'number | null'.
This commit fixes this.

The behavior of the mentioned components remains exactly the same, no
new modifications were added.