feat(React): add a page for searching for products
This commit adds a new page to the React application (/search/products)
that allows users to search for products using a search query. This page
is available to both authenticated and anonymous users. Additionally,
pagination is used to ensure that products are shown by page and not all
at once.

As of now, no autosuggestions are shown, because this functionality is
not implemented in the backend app.