feat(API migrations): drop 'status_id' from the 'orders' table
In light of the recent addition of the 'order_status_history' table,
it's now essential to drop the 'status_id' column from the 'orders'
table. This column is redundant because order statuses are now managed
within the 'order_status_history' table. To retrieve the latest order
status, querying the newly created table for the most recent status
associated with a specific order is all that's needed.

Furthermore, the order creation endpoint (POST /api/user/order) has been
updated to reflect this new change. Upon order creation, a new row is
added to the 'order_status_history' table, indicating the creation of a
new order.