feat(API seeds): add sample orders
This commit introduces 4 sample orders for testing purposes using the
'db-seeds/14_orders.ts' file. This file prepopulates three tables:
- orders (stores orders)
- orders_products (stores products that are associated with orders)
- order_status_history (tracks the history of order statuses)
It was decided to prepopulates these tables in one file instead of
creating a separate file for each table, because these DB tables
are tightly coupled and are all closely related to the concept of an
order. Holding all of this logic inside one file will make it easier to
make similar changes in the future.