feat(API): add an endpoint for fetching a list of orders
This commit introduces a new "GET /api/user/orders" endpoint, which is
used for retrieving a list of orders for the user who initiated the
request. This endpoint will be used for displaying a history of orders
in the React application.

For more details about this endpoint please refer to the API docs.

Note:
This endpoint has a limitation, which is that the returned orders are
not paginated, meaning the API retrieves every order that the user has
ever made, which is not scalable in the long run. This behavior will be
fixed in the future.