fix(API seeding): insert random delivery warehouse when creating orders
Seed files are a way to prepopulate database tables. They're mainly used
for testing purposes.

There was a problem with the '14_orders.ts' seed file, which is
responsible for creating orders. This seed file tried to create orders
that are tied to non-existent postal delivery warehouses. The root of
the problem is that the warehouse names were hardcoded, and Nova Poshta
changed some of their warehouses' names. Since these orders were created
only for testing purposes, it was decided to configure the seed file
to choose random delivery warehouses for the created orders. This way,
the seed file will always insert a correct delivery warehouse for the
given order.