feat(API migrations): add 'service_centers' table
This commit introduces a new DB table to manage service centers. When a
user experiences issues with a purchased product within the warranty
period, they can visit our nearest offline shop. An administrator can
then add a warranty request to the database and transfer the product to
the appropriate service center for diagnostics and repair.

The 'service_centers' table includes the following fields:
- id
- address_id (specifying the address of the service center)

These fields are utilized to construct a name for each service center,
such as "Service Center No. 3 (16 Heroes of Independence St., Zhytomyr,
03003)".